Service Cleanup, DB Additions, Image Processing
-Added a new image processing service that strips metadata and resizes thumbnails
-Added isPublic and userID to all buildData DB tables to increase DB lookup efficiency
-Updated services to reflect the DB changes
-Moved token checking into buildData services to utilize the new fields
-Added some checks for a "local" Environement variable to aid in local debugging

+++ Server/BuildBazaar/BuildBazaarCore/Services/ReferenceImageService.cs
@@ -5,6 +5,7 @@ using Org.BouncyCastle.Asn1.Ocsp;
 using System.ComponentModel.Design;
 using System.Data.Common;
 using System.IdentityModel.Tokens.Jwt;
+using System.IO;
 using System.Security.Claims;
 
 namespace BuildBazaarCore.Services
@@ -12,7 +13,7 @@ namespace BuildBazaarCore.Services
     public interface IReferenceImageService
     {
         Task<IActionResult> UploadReferenceImage(IFormFile file, int buildID, JwtSecurityToken token);
-        Task<IActionResult> GetReferenceImages(uint buildID, bool isPublic, JwtSecurityToken token);
+        Task<IActionResult> GetReferenceImages(uint buildID, JwtSecurityToken token);
         Task<IActionResult> SaveImageOrder(ImageOrderRequest request);
         Task<IActionResult> DeleteReferenceImage(uint imageID, JwtSecurityToken token);
     }
@@ -21,17 +22,19 @@ namespace BuildBazaarCore.Services
         private readonly IConfigService _configService;
         private readonly IBuildService _buildService;
         private readonly IAwsService _awsService;
-        public ReferenceImageService(IConfigService configService, IBuildService buildService, IAwsService awsService)
+        private readonly IImageProcessingService _imageProcessingService;
+        public ReferenceImageService(IConfigService configService, IBuildService buildService, IAwsService awsService, IImageProcessingService imageProcessingService)
         {
             _configService = configService;
             _buildService = buildService;
             _awsService = awsService;
+            _imageProcessingService = imageProcessingService;
         }
+
         public async Task<IActionResult> UploadReferenceImage(IFormFile file, int buildID, JwtSecurityToken token)
         {
             using (MySqlConnection connection = new MySqlConnection(_configService.GetConnectionString()))
             {
-                var userIDClaim = token.Claims.FirstOrDefault(c => c.Type == ClaimTypes.NameIdentifier)?.Value;
                 await connection.OpenAsync();
 
                 using (var transaction = connection.BeginTransaction())
@@ -39,9 +42,32 @@ namespace BuildBazaarCore.Services
                     var s3ImageFilePath = "";
                     try
                     {
-                        
+                        uint userIDClaim = getUserIDFromToken(token);
+                        bool isPublic = false;
+
+                        string buildQuery = "SELECT UserID, isPublic FROM Builds WHERE BuildID = @buildID";
+                        using (MySqlCommand command = new MySqlCommand(buildQuery, connection))
+                        {
+                            command.Parameters.AddWithValue("@buildID", buildID);
+                            using (DbDataReader reader = await command.ExecuteReaderAsync())
+                            {
+                                if (!await reader.ReadAsync())
+                                {
+                                    return Json(new { success = false, errorMessage = "Build not found." });
+                                }
+
+                                uint userID = Convert.ToUInt32(reader["UserID"]);
+                                isPublic = Convert.ToBoolean(reader["isPublic"]);
+
+                                if (userID != userIDClaim)
+                                {
+                                    return Json(new { success = false, errorMessage = "Unauthorized access." });
+                                }
+                            }
+                        }
+
                         var imageFileName = Guid.NewGuid().ToString() + Path.GetExtension(file.FileName);
-                        s3ImageFilePath = Path.Combine("Users", userIDClaim, buildID.ToString(), imageFileName);
+                        s3ImageFilePath = Path.Combine("Users", userIDClaim.ToString(), buildID.ToString(), imageFileName);
 
                         int maxOrder;
                         var maxOrderQuery = "SELECT IFNULL(MAX(imageOrder), 0) FROM Images WHERE buildID = @buildID";
@@ -51,59 +77,73 @@ namespace BuildBazaarCore.Services
                             maxOrder = Convert.ToInt32(command.ExecuteScalar());
                         }
 
-                        var imageQuery = "INSERT INTO Images (filePath, imageOrder, typeID, buildID, userID) VALUES (@filePath, @imageOrder, 2, @buildID, @userID)";
+                        var imageQuery = "INSERT INTO Images (filePath, imageOrder, typeID, buildID, userID, isPublic) VALUES (@filePath, @imageOrder, 2, @buildID, @userID, isPublic)";
                         using (MySqlCommand command = new MySqlCommand(imageQuery, connection, transaction))
                         {
                             command.Parameters.AddWithValue("@filePath", s3ImageFilePath);
+                            command.Parameters.AddWithValue("@imageOrder", maxOrder + 1);
                             command.Parameters.AddWithValue("@buildID", buildID);
                             command.Parameters.AddWithValue("@userID", userIDClaim);
-                            command.Parameters.AddWithValue("@imageOrder", maxOrder + 1);
+                            command.Parameters.AddWithValue("@isPublic", isPublic);
+                            
                             command.ExecuteNonQuery();
                         }
 
-                        using (var memoryStream = new MemoryStream())
+                        using (var memoryStream = await _imageProcessingService.StripExifData(file))
                         {
-                            await file.CopyToAsync(memoryStream);
-                            memoryStream.Position = 0; // Reset stream position after copy
-                            await _awsService.UploadStreamToS3(memoryStream, s3ImageFilePath);
+                            if (_configService.GetEnvironment() != "local")
+                            {
+                                // Upload the clean image to S3
+                                await _awsService.UploadStreamToS3(memoryStream, s3ImageFilePath);
+                            }
+                            else
+                            {
+                                using (FileStream localFile = new FileStream("localImages/" + imageFileName, FileMode.Create))
+                                {
+                                    memoryStream.CopyTo(localFile);
+                                }
+                            }
                         }
+
                         await transaction.CommitAsync();
                         return Json(new { success = true });
                     }
                     catch (Exception ex)
                     {
                         await transaction.RollbackAsync();
-                        await _awsService.DeleteFileFromS3(s3ImageFilePath);
+                        if (_configService.GetEnvironment() != "local")
+                        {
+                            await _awsService.DeleteFileFromS3(s3ImageFilePath);
+                        }
                         return Json(new { success = false, errorMessage = ex.Message });
                     }
                 }
-
             }
         }
 
-        public async Task<IActionResult> GetReferenceImages(uint buildID, bool isPublic, JwtSecurityToken token)
+        public async Task<IActionResult> GetReferenceImages(uint buildID, JwtSecurityToken token)
         {
-            var (success, validBuildID, errorMessage) = await _buildService.getBuildID(buildID, isPublic, token);
-
-            if (!success)
-            {
-                return Json(new { success = false, errorMessage });
-            }
-
             try
             {
                 List<ImageModel> images = new List<ImageModel>();
+                uint userIDClaim = getUserIDFromToken(token);
+
                 using (MySqlConnection connection = new MySqlConnection(_configService.GetConnectionString()))
                 {
                     await connection.OpenAsync();
                     string query = "SELECT * FROM Images WHERE Images.typeID = 2 AND buildID = @buildID";
                     using (MySqlCommand command = new MySqlCommand(query, connection))
                     {
-                        command.Parameters.AddWithValue("@buildID", validBuildID);
+                        command.Parameters.AddWithValue("@buildID", buildID);
                         using (DbDataReader reader = await command.ExecuteReaderAsync())
                         {
                             while (await reader.ReadAsync())
                             {
+                                bool isPublic = Convert.ToBoolean(reader["isPublic"]);
+                                if (!isPublic && userIDClaim != Convert.ToUInt32(reader["userID"]))
+                                {
+                                    continue;
+                                }
                                 ImageModel image = new ImageModel();
                                 image.imageID = Convert.ToUInt32(reader["imageID"]);
                                 image.buildID = Convert.ToUInt32(reader["buildID"]);
@@ -152,14 +192,13 @@ namespace BuildBazaarCore.Services
             }
         }
 
-
         public async Task<IActionResult> DeleteReferenceImage(uint imageID, JwtSecurityToken token)
         {
             try
             {
                 using (MySqlConnection connection = new MySqlConnection(_configService.GetConnectionString()))
                 {
-                    var userIDClaim = token.Claims.FirstOrDefault(c => c.Type == ClaimTypes.NameIdentifier)?.Value;
+                    uint userIDClaim = getUserIDFromToken(token);
                     string filePath;
                     await connection.OpenAsync();
 
@@ -177,7 +216,13 @@ namespace BuildBazaarCore.Services
                     {
                         return Json(new { success = false, errorMessage = "File not found" });
                     }
-                    bool isDeleted = await _awsService.DeleteFileFromS3(filePath);
+
+                    bool isDeleted = true;
+                    if (_configService.GetEnvironment() != "local")
+                    {
+                        isDeleted = await _awsService.DeleteFileFromS3(filePath);
+                    }
+
                     if (!isDeleted)
                     {
                         return Json(new { success = false, errorMessage = "Error deleting file" });
@@ -189,12 +234,8 @@ namespace BuildBazaarCore.Services
                         command.Parameters.AddWithValue("@imageID", imageID);
                         command.Parameters.AddWithValue("@userID", userIDClaim);
 
-                        int rowsAffected = await command.ExecuteNonQueryAsync();
-                        if (rowsAffected <= 0)
-                        {
-                            return Json(new { success = false, errorMessage = "No records deleted." });
+                        await command.ExecuteNonQueryAsync();
 
-                        }
                         return Json(new { success = true });
                     }
                 }
